Travis: upload build artifacts
