+++ scripts/build.sh
@@ -87,7 +87,7 @@ buildcoordinator() {
     fi
 
     printf "\nCompiling..."
-    go build -o build/coordinator.exe ./src/peach_client_coordinator || fail
+    go build -o build/coordinator-$version.exe ./src/peach_client_coordinator || fail
     printf "\nDone building client coordinator\n"
     storehash "peach_client_coordinator"
 }
@@ -104,20 +104,12 @@ builddiscordclient() {
     fi
 
     printf "\nCompiling..."
-    version=$(git describe --tags)
-    branch=$(git rev-parse --abbrev-ref HEAD)
-    if [[ "$branch" == "master" ]]; then
-        branch=""
-    else
-        branch="-$branch"
-    fi
-    version=${version%-*-*}
-    version="$version$branch"
+    
     printf "package main\n
 // VERSION represents the release the bot is currently running on.
 // This file is automatically generated by the build script. Do not edit.
 const VERSION = \"$version\"" > src/peach_discord_client/version.go
-    go build -o build/discordclient.exe ./src/peach_discord_client || fail
+    go build -o build/discordclient-$version.exe ./src/peach_discord_client || fail
     printf "\nDone building discord client\n"
     storehash "peach_discord_client"
 }
@@ -134,7 +126,13 @@ buildlauncher() {
     fi
 
     printf "\nCompiling..."
-    go build -o build/launcher.exe ./src/peach_launcher || fail
+
+    printf "package main\n
+// VERSION represents the release the bot is currently running on.
+// This file is automatically generated by the build script. Do not edit.
+const VERSION = \"$version\"" > src/peach_launcher/version.go
+
+    go build -o build/launcher-$version.exe ./src/peach_launcher || fail
     printf "\nDone building launcher\n"
     storehash "peach_launcher"
 }
@@ -153,4 +151,15 @@ then
         exit
     fi
 fi
+
+version=$(git describe --tags)
+branch=$(git rev-parse --abbrev-ref HEAD)
+if [[ "$branch" == "master" ]]; then
+    branch=""
+else
+    branch="-$branch"
+fi
+version=${version%-*-*}
+version="$version$branch"
+
 build
